requirements.txt
torch==2.6.0
torchvision==0.21.0
mlflow==2.13.0
kagglehub==0.3.4
Pillow==10.3.0
scikit-learn==1.4.2
matplotlib==3.8.4
seaborn==0.13.2
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
pytest==8.2.0
requests==2.31.0
numpy==1.26.4
